movie_service/requirements.txt
Django==1.8
djangorestframework==3.1.1
Pygments==2.0.2
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
